+++ thedozens/API/admin.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
-from django.contrib import admin
 from API.models import Insult, InsultReview
+from django.contrib import admin
 
 all_models = [Insult, InsultReview]
 for model in all_models:
