+++ thedozens/API/models.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+import datetime
 import os
 
 from django.conf import settings
@@ -7,7 +8,6 @@ from django.db import models
 from django.utils.translation import gettext_lazy as _
 from logtail import LogtailHandler
 from loguru import logger
-import datetime
 
 NOW = datetime.datetime.now()
