+++ thedozens/graphQL/schema.py
@@ -1,8 +1,7 @@
 # -*- coding: utf-8 -*-
-from graphQL.query import Query
-from graphQL.mutations import Mutation
-from graphQL.subscriptions import Subscription
 from graphene import Schema
-
+from graphQL.mutations import Mutation
+from graphQL.query import Query
+from graphQL.subscriptions import Subscription
 
 schema = Schema(query=Query)  # , mutation=Mutation, subscription=Subscription)
