+++ thedozens/thedozens/settings.py
@@ -6,19 +6,22 @@ Generated by 'django-admin startproject' using Django 4.2.6.
 
 
 """
-from pathlib import Path
 import os
+import sys
 from datetime import datetime
-from django.core.mail import send_mail
-from datetime import datetime
-from loguru import logger
+from pathlib import Path
+
 from logtail import LogtailHandler
+from loguru import logger
 
 GLOBAL_NOW = datetime.now()
+
+LOGGING_CONFIG = None 
 
 
 # SECTION - Application definition
 ROOT_URLCONF = "thedozens.urls"
+SITE_ID = 1
 WSGI_APPLICATION = "thedozens.wsgi.application"
 BASE_DIR = Path(__file__).resolve().parent.parent
 SECRET_KEY = os.environ["SECRET_KEY"]
@@ -31,6 +34,7 @@ INSTALLED_APPS = [
     "django.contrib.auth",
     "django.contrib.contenttypes",
     "django.contrib.sessions",
+    "django.contrib.sites",
     "django.contrib.messages",
     "django.contrib.staticfiles",
     # Third-Party Apps
@@ -40,9 +44,11 @@ INSTALLED_APPS = [
     "debug_toolbar",
     "graphene_django",
         'djoser',
+            'djoser.webauthn',
     "crispy_forms",
     "crispy_bootstrap5",
     "django_prometheus",
+    "drf_spectacular",
     # Project Apps
     "API",
     "graphQL",
@@ -52,6 +58,8 @@ TIME_ZONE = "America/Chicago"
 USE_I18N = True
 USE_TZ = True
 MIDDLEWARE = [
+    
+    "kolo.middleware.KoloMiddleware",
         "django.middleware.cache.UpdateCacheMiddleware",
     "django.middleware.security.SecurityMiddleware",
     "django.contrib.sessions.middleware.SessionMiddleware",
@@ -87,9 +95,7 @@ CACHES = {
     "default": {
         "BACKEND": "django.core.cache.backends.redis.RedisCache",
         "LOCATION": os.environ["REDIS_CACHE_URI"],
-        "OPTIONS": {
-            "CLIENT_CLASS": "django_redis.client.DefaultClient",
-        },
+
     }
 }
 CACHE_MIDDLEWARE_ALIAS = "default"
@@ -114,12 +120,6 @@ AUTH_PASSWORD_VALIDATORS = [
         "NAME": "django.contrib.auth.password_validation.NumericPasswordValidator",
     },
 ]
-AUTHENTICATION_BACKENDS = [
-     # Needed to login by username in Django admin, regardless of `allauth`
-    'django.contrib.auth.backends.ModelBackend',
-     # `allauth` specific authentication methods, such as login by email
-    'allauth.account.auth_backends.AuthenticationBackend',
- ]
 
 #!SECTION
 
@@ -220,11 +220,15 @@ CRISPY_ALLOWED_TEMPLATE_PACKS = "bootstrap5"
 CRISPY_TEMPLATE_PACK = "bootstrap5"
 # !SECTION
 
-PRIMARY_LOG_FILE = os.path.join( BASE_DIR, "standup", "logs", "primary_ops.log")
-CRITICAL_LOG_FILE = os.path.join( BASE_DIR, "standup", "logs", "fatal.log")
-DEBUG_LOG_FILE = os.path.join( BASE_DIR, "standup", "logs", "utility.log")
+PRIMARY_LOG_FILE = os.path.join( BASE_DIR, "logs", "primary_ops.log")
+CRITICAL_LOG_FILE = os.path.join( BASE_DIR, "logs", "fatal.log")
+DEBUG_LOG_FILE = os.path.join( BASE_DIR, "logs", "utility.log")
 LOGTAIL_HANDLER = LogtailHandler(source_token=os.environ["LOGTAIL_API_KEY"])
+DEFAULT_HANDLER = sys.stdout
 
 logger.add(DEBUG_LOG_FILE, diagnose=True, catch=True, backtrace=True, level="DEBUG")
+logger.add(DEFAULT_HANDLER, diagnose=True, catch=True, backtrace=True, level="DEBUG")
 logger.add(PRIMARY_LOG_FILE, diagnose=False, catch=True, backtrace=False, level="INFO")
 logger.add(LOGTAIL_HANDLER, diagnose=False, catch=True, backtrace=False, level="INFO")
+
+
