--- thedozens/thedozens/test_settings.py
@@ -1,126 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Django TEST settings for thedozens project.
-
-Generated by 'django-admin startproject' using Django 4.2.6.
-
-
-"""
-import os
-from pathlib import Path
-
-from dotenv import load_dotenv
-
-load_dotenv()
-
-
-# SECTION - Application definition
-BASE_DIR = Path(__file__).resolve().parent.parent
-SECRET_KEY = os.getenv("SECRET_KEY")
-DEBUG = True
-ALLOWED_HOSTS = ["*"]
-INSTALLED_APPS = [
-    # Django-Installed Apps
-    "django.contrib.admin",
-    "django.contrib.auth",
-    "django.contrib.contenttypes",
-    "django.contrib.sessions",
-    "django.contrib.messages",
-    "django.contrib.staticfiles",
-    # Third-Party Apps
-    "restframework",
-    "debug_toolbar",
-    # Project Apps
-    "API",
-]
-LANGUAGE_CODE = "en-us"
-TIME_ZONE = "America/New_York"
-USE_I18N = True
-USE_TZ = True
-MIDDLEWARE = [
-    "django.middleware.security.SecurityMiddleware",
-    "django.contrib.sessions.middleware.SessionMiddleware",
-    "django.middleware.common.CommonMiddleware",
-    "django.middleware.csrf.CsrfViewMiddleware",
-    "django.contrib.auth.middleware.AuthenticationMiddleware",
-    "django.contrib.messages.middleware.MessageMiddleware",
-    "debug_toolbar.middleware.DebugToolbarMiddleware",
-    "django.middleware.clickjacking.XFrameOptionsMiddleware",
-]
-
-ROOT_URLCONF = "thedozens.urls"
-
-
-WSGI_APPLICATION = "thedozens.wsgi.application"
-#!SECTION
-
-# SECTION - Database and Caching
-DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
-DATABASES = {
-    "default": {
-        "ENGINE": "django.db.backends.postgresql_psycopg2",
-        "NAME": os.getenv("DB_NAME"),
-        "USER": os.getenv("DB_USER"),
-        "PASSWORD": os.getenv("DB_PASSWORD"),
-        "HOST": os.getenv("PG_DATABASE_HOST"),
-        "PORT": os.getenv("DB_PORT"),
-        "OPTIONS": {"ssl": True},
-    }
-}
-
-CACHES = {
-    "default": {
-        "BACKEND": "django.core.cache.backends.redis.RedisCache",
-        "LOCATION": os.getenv("REDIS_CACHE_URI"),
-    }
-}
-#!SECTION
-
-# SECTION - Password validatio and User Authentication
-
-AUTH_PASSWORD_VALIDATORS = [
-    {
-        "NAME": "django.contrib.auth.password_validation.UserAttributeSimilarityValidator",
-    },
-    {
-        "NAME": "django.contrib.auth.password_validation.MinimumLengthValidator",
-    },
-    {
-        "NAME": "django.contrib.auth.password_validation.CommonPasswordValidator",
-    },
-    {
-        "NAME": "django.contrib.auth.password_validation.NumericPasswordValidator",
-    },
-]
-
-#!SECTION
-
-
-# SECTION - Static files & Templatea
-
-TEMPLATES = [
-    {
-        "BACKEND": "django.template.backends.django.DjangoTemplates",
-        "DIRS": [],
-        "APP_DIRS": True,
-        "OPTIONS": {
-            "context_processors": [
-                "django.template.context_processors.debug",
-                "django.template.context_processors.request",
-                "django.contrib.auth.context_processors.auth",
-                "django.contrib.messages.context_processors.messages",
-            ],
-        },
-    },
-]
-
-STATIC_URL = "static/"
-#!SECTION
-
-# SECTION - DRF Settings
-REST_FRAMEWORK = {
-    "DEFAULT_PERMISSION_CLASSES": [
-        "rest_framework.permissions.DjangoModelPermissionsOrAnonReadOnly"
-    ]
-}
-#!SECTION
